+++ src/tiled/scriptmodule.h
@@ -20,9 +20,10 @@
 
 #pragma once
 
-#include "documentmanager.h"
+#include "compression.h"
 #include "id.h"
 #include "issuesdock.h"
+#include "properties.h"
 
 #include <QJSValue>
 #include <QObject>
@@ -34,6 +35,7 @@ class QAction;
 
 namespace Tiled {
 
+class Document;
 class EditableAsset;
 class MapEditor;
 class ScriptImage;
@@ -121,6 +123,19 @@ public:
     Q_INVOKABLE Tiled::ScriptTilesetFormatWrapper *tilesetFormatForFile(const QString &fileName) const;
 
     Q_INVOKABLE void extendMenu(const QByteArray &idName, QJSValue items);
+
+    // Synchronized with Tiled::CompressionMethod
+    enum CompressionMethod {
+        Gzip,
+        Zlib,
+        Zstandard
+    };
+    Q_ENUM(CompressionMethod)
+
+    Q_INVOKABLE QByteArray compress(const QByteArray &data, CompressionMethod method = Zlib, int compressionLevel = -1);
+    Q_INVOKABLE QByteArray decompress(const QByteArray &data, CompressionMethod method = Zlib);
+    Q_INVOKABLE QByteArray toBase64(const QByteArray &data);
+    Q_INVOKABLE QByteArray fromBase64(const QByteArray &data);
 
 signals:
     void assetCreated(Tiled::EditableAsset *asset);